feat: 수업 일지 상세 조회 api 연동
